requirements.txt
celery==5.3.4
redis==5.0.1
click==8.1.7
pandas==2.1.4
pyarrow==14.0.1
requests==2.31.0
python-dotenv==1.0.0
polygon-api-client==1.10.1
tenacity==8.2.3
pydantic==2.5.2
boto3==1.34.0
flower==2.0.1
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==11.0.3
python-multipart==0.0.6